W5500 driver register addressing
Fixed 16/32-bit addressing issues when executing block send commands.
Connection over TCP appear to function, transmit sends garbage.  This
seems related to the TxWR pointer being corrupted when reading, however,
this could be normal operation of the device, since TCP connection
auto-increments the packet buffer pointers after connection has been
established.

+++ E2ForLife_W5x00.cylib/E2ForLife_W5x00_v2_0/API/W5x00.c
@@ -1003,10 +1003,10 @@ uint16 `$INSTANCE_NAME`_GetRxSize( uint8 socket )
  */
 void `$INSTANCE_NAME`_ProcessTxData(uint8 socket, uint16 offset, uint8* buffer, uint16 length)
 {
-	uint16 addr;
-	uint16 base;
-	uint16 PointerOffset;
-	uint16 size;
+	uint32 addr;
+	uint32 base;
+	uint32 PointerOffset;
+	uint32 size;
 	
 	/*
 	 * Read the offset pointer, and calculate the base address for the start of write
@@ -1059,10 +1059,10 @@ void `$INSTANCE_NAME`_ProcessTxData(uint8 socket, uint16 offset, uint8* buffer, 
  */
 void `$INSTANCE_NAME`_ProcessRxData(uint8 socket, uint16 offset, uint8* buffer, uint16 length, uint8 flags)
 {
-	uint16 addr;
-	uint16 base;
-	uint16 PointerOffset;
-	uint16 size;
+	uint32 addr;
+	uint32 base;
+	uint32 PointerOffset;
+	uint32 size;
 	
 	/*
 	 * Read the offset pointer, and calculate the base address for the start of read
